+++ src/dmcigd/core/objects/interfaces/SolidObject.java
@@ -6,7 +6,7 @@ public interface SolidObject extends GameObject {
 	
 	public boolean isDestroyed();
 	
-	public void onPush(Entity entity, int v);
+	public void onPush(Entity entity, float v);
 	public boolean onAttack(int damage, boolean flipped);
 	
 }
